--- node.go
@@ -1,40 +1,0 @@
-// Copyright (c) 2017 Che Wei, Lin
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package main
-
-import "github.com/John-Lin/ovsdb"
-
-type Node interface {
-	addPort(ifName string) error
-	setCtrl(hostport string) error
-	setupVeth(sandbox string, ifName string, mtu int) (*Host, error)
-	setIfaceIP(sandbox, ifName, address string) error
-}
-
-// OVSSwitch is a bridge instance
-type OVSSwitch struct {
-	nodeType     string
-	bridgeName   string
-	ctrlHostPort string
-	ovsdb        *ovsdb.OvsDriver
-}
-
-// Host is a host instance
-type Host struct {
-	nodeType string
-	name     string
-	ifName   string
-	sandbox  string
-}
